requirements.txt
google-generativeai
python-dotenv
pandas
Pillow
torch
transformers
huggingface_hub
timm